Rename and move trait to assert array content
