+++ tests/Integration/Api/V1/TeamsTest.php
@@ -5,11 +5,11 @@ namespace Tests\Integration\Api\V1;
 use App\Models\Division;
 use App\Models\Venue;
 use Tests\ApiTestCase;
-use Tests\AssertArrayContent;
+use Tests\Concerns\InteractsWithArrays;
 
 class TeamsTest extends ApiTestCase
 {
-    use AssertArrayContent;
+    use InteractsWithArrays;
 
     public function testGettingAllTeams(): void
     {
