+++ tests/Unit/Resources/FixtureResourceTest.php
@@ -8,12 +8,12 @@ use App\Models\Fixture;
 use App\Models\Venue;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\Request;
-use Tests\AssertArrayContent;
+use Tests\Concerns\InteractsWithArrays;
 use Tests\TestCase;
 
 class FixtureResourceTest extends TestCase
 {
-    use RefreshDatabase, AssertArrayContent;
+    use RefreshDatabase, InteractsWithArrays;
 
     private $division;
     private $homeTeam;
